backend/requirements.txt
# Install with: pip install -r requirements.txt --index-url https://pypi.org/simple
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
wheel==0.42.0
setuptools==68.2.2 